requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart

# Data Validation
pydantic
pydantic-core
email-validator

# AI Integration
groq
# openai  # Alternative AI provider (optional)

# File Processing
pypdf

# Environment Management
python-dotenv

# HTTP Client (for AI APIs)
httpx
httpcore

# Utilities
typing-extensions
annotated-types
anyio
